dashboard: c-channel nudge controls; move incidents to settings; drop bins card, logs tab
- CameraFeed gains an optional headerActions snippet; the three split-feeder
  views get right-justified -1 deg / +1 deg / 180 deg CW buttons (motor
  degrees via 130/12 gear ratio, stored per-stepper speed, one in-flight
  action, Popover tooltips).
- Incident handling extracted to IncidentHandlingSection under the Models
  settings page; dashboard block removed.
- Dashboard bins card removed (SortingStatusCard deleted); Logs tab removed
  from the top bar.
