+++ software/sorter/frontend/src/lib/components/AppHeader.svelte
@@ -385,16 +385,6 @@
 					Records
 				</a>
 				<a
-					href="/logs"
-					class="px-3 py-1.5 text-sm font-medium transition-colors {page.url.pathname.startsWith(
-						'/logs'
-					)
-						? 'border-b-2 border-primary text-primary'
-						: 'text-text-muted hover:bg-bg hover:text-text'}"
-				>
-					Logs
-				</a>
-				<a
 					href="/settings"
 					class="px-3 py-1.5 text-sm font-medium transition-colors {page.url.pathname.startsWith(
 						'/settings'
