+++ software/sorter/frontend/src/routes/settings/hive/models/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import HiveModelsSection from '$lib/components/settings/HiveModelsSection.svelte';
 	import HiveRuntimesSection from '$lib/components/settings/HiveRuntimesSection.svelte';
+	import IncidentHandlingSection from '$lib/components/settings/IncidentHandlingSection.svelte';
 	import SectionCard from '$lib/components/settings/SectionCard.svelte';
 </script>
 
@@ -23,5 +24,12 @@
 		description="For each shipping model format, which inference backends this machine can run and how fast they're expected to be on the selected model."
 	>
 		<HiveRuntimesSection />
+	</SectionCard>
+
+	<SectionCard
+		title="Incident Handling"
+		description="How the machine responds when something goes wrong mid-sort: ignore the condition, wait for an operator, or recover automatically where supported."
+	>
+		<IncidentHandlingSection />
 	</SectionCard>
 </div>
